requirements.txt
anthropic>=0.18.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pymongo
openai>=1.0.0